Move `validateLocation` and `processValidation` to `src/geoclient.js`
